requirements.txt
#Completar con todas las librerias utilizadas en el programa y sus versiones

PySimpleGUI == 4.60.4
pillow == 9.5.0
pandas == 2.0.2
matplotlib == 3.7.1
worldcloud == 1.9.2
numpy == 1.24.3
tabulate == 0.9.0
streamlit == 1.22.0